+++ vm/address/src/lib.rs
@@ -3,16 +3,16 @@
 
 mod errors;
 mod network;
+mod payload;
 mod protocol;
 pub use self::errors::Error;
 pub use self::network::Network;
-pub use self::protocol::{Payload, Protocol};
+pub use self::payload::Payload;
+pub use self::protocol::Protocol;
 
 use data_encoding::Encoding;
 use data_encoding_macro::{internal_new_encoding, new_encoding};
-use encoding::{
-    blake2b_variable, de, ser, serde_bytes, Cbor, CodecProtocol, Error as EncodingError,
-};
+use encoding::{blake2b_variable, de, ser, serde_bytes, Cbor};
 use std::fmt;
 use std::hash::Hash;
 use std::str::FromStr;
@@ -29,7 +29,6 @@ pub const CHECKSUM_HASH_LEN: usize = 4;
 const MAX_ADDRESS_LEN: usize = 84 + 2;
 const MAINNET_PREFIX: &str = "f";
 const TESTNET_PREFIX: &str = "t";
-const BUFFER_SIZE: usize = 1024;
 
 // TODO pull network from config (probably)
 const NETWORK_DEFAULT: Network = Network::Testnet;
@@ -104,7 +103,7 @@ impl Address {
     }
 
     /// Returns data payload of Address
-    pub fn payload(&self) -> Vec<u8> {
+    pub fn payload_bytes(&self) -> Vec<u8> {
         self.payload.to_bytes()
     }
 
@@ -121,7 +120,7 @@ impl Address {
 
     /// Returns encoded bytes of Address
     pub fn to_bytes(&self) -> Vec<u8> {
-        let mut bz: Vec<u8> = self.payload().to_vec();
+        let mut bz: Vec<u8> = self.payload_bytes();
         bz.insert(0, self.protocol() as u8);
         bz
     }
@@ -227,21 +226,12 @@ impl<'de> de::Deserialize<'de> for Address {
 
 impl Cbor for Address {}
 
-impl From<Error> for EncodingError {
-    fn from(err: Error) -> EncodingError {
-        EncodingError::Marshalling {
-            description: err.to_string(),
-            protocol: CodecProtocol::Cbor,
-        }
-    }
-}
-
 /// encode converts the address into a string
 fn encode(addr: &Address) -> String {
     match addr.protocol() {
         Protocol::Secp256k1 | Protocol::Actor | Protocol::BLS => {
             let ingest = addr.to_bytes();
-            let mut bz = addr.payload().to_vec();
+            let mut bz = addr.payload_bytes();
 
             // payload bytes followed by calculated checksum
             bz.extend(checksum(&ingest));
@@ -252,17 +242,12 @@ fn encode(addr: &Address) -> String {
                 ADDRESS_ENCODER.encode(bz.as_mut()),
             )
         }
-        Protocol::ID => {
-            let mut buf = [0; BUFFER_SIZE];
-            buf[..addr.payload().len()].copy_from_slice(&addr.payload());
-            let mut readable = &buf[..];
-            format!(
-                "{}{}{}",
-                addr.network.to_prefix(),
-                addr.protocol().to_string(),
-                leb128::read::unsigned(&mut readable).expect("should read encoded bytes"),
-            )
-        }
+        Protocol::ID => format!(
+            "{}{}{}",
+            addr.network.to_prefix(),
+            addr.protocol().to_string(),
+            from_leb_bytes(&addr.payload_bytes()).expect("should read encoded bytes"),
+        ),
     }
 }
